requirements.txt
numpy==1.24.1
opencv_contrib_python==4.8.0.76
opencv_python==4.8.0.76
Pillow==10.2.0
pyfvvdp==1.2.0
scikit_learn==1.3.0
scikit_video==1.1.11
scipy==1.12.0
tabulate==0.9.0
torch==2.0.1+cu118
torchsummary==1.5.1
torchvision==0.15.2+cu118
